api/merging/requirements.txt
numpy
sentence-transformers
faiss-cpu
nltk